requirements.txt
bcrypt
fastapi
sqlalchemy
uvicorn
PyJWT